requirements.txt
# Core Agent Framework
uagents==0.12.0
uagents-ai-engine==0.4.0

# MeTTa/Hyperon
hyperon>=0.2.8

# Web3 & Blockchain
web3==6.11.0
cosmpy==0.9.0

# HTTP & Async
aiohttp==3.9.0
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Cryptocurrency APIs
ccxt==4.5.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.0

# Utilities
python-dateutil==2.8.2